Added Monarch Butterfly model
Butterfly model added to butterfly room. I can't seem to figure out how to get the underside of its wings visible. Files for this are stored in the Butterfly_Model folder in Assets. There are additional files for this model directly in the Assets folder that I am unable to move to the Butterfly_Model folder.